+++ dces-drc-integration/build.gradle
@@ -164,7 +164,7 @@ sonar {
 		property "sonar.projectKey", "ministryofjustice_laa-dces-drc-integration"
 
 		property "sonar.exclusions", "**/entity/**.java , **/model/**.java , **/dto/**.java , **/config/**.java, **/jms/**.java,**/exception/**.java,**/handler/**.java,**/maatapi/**Client**.java,**/generated/**, **/listener/stub/DrcStubRestController.java"
-		property "sonar.coverage.exclusions", "**/DcesDrcIntegrationApplication.java, **/listener/stub/DrcStubRestController.java"
+		property "sonar.coverage.exclusions", "**/DcesDrcIntegrationApplication.java, **/controller/DrcStubRestController.java"
 		property "sonar.coverage.jacoco.xmlReportPaths",
 				"${project.buildDir}/reports/jacoco/test/jacocoTestReport.xml"
 	}
