+++ dces-drc-integration/src/main/java/uk/gov/justice/laa/crime/dces/integration/controller/DcesDrcErrorController.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.rest.api.controller;
+package uk.gov.justice.laa.crime.dces.integration.controller;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.request.WebRequest;
 import uk.gov.justice.laa.crime.dces.integration.rest.common.ErrorResponse;
-import uk.gov.justice.laa.crime.dces.integration.tracing.TraceService;
+import uk.gov.justice.laa.crime.dces.integration.service.TraceService;
 
 import java.util.Map;
